requirements.txt
django<1.9
psycopg2
djangorestframework
django-pipeline
jsmin
